+++ UltimateCastbars/Backend/Core/SemiDynamic/Normal.lua
@@ -48,6 +48,7 @@ function CASTBAR_API:OnUnitSpellcastStop(unit, castGUID, spellID, castBarID)
         bar.flags.prevType = nil
         bar.current_spellID = nil
         bar._ucbUnit, bar._ucbCfg, bar._ucbCastType, bar._ucbVars,  bar._ucbSpellID = nil, nil, nil, nil, nil
+        if bar.effects.spark then bar.effects.spark:Hide() end
         return true
      end
     return false